Module3/cdc-pipeline/requirements.txt
psycopg2-binary==2.9.9
confluent-kafka==2.3.0
requests==2.31.0
rich==13.7.1
